+++ cassandra-migration-spring-boot-starter/src/test/java/org/cognitor/cassandra/migration/spring/CassandraMigrationAutoConfigurationTest.java
@@ -52,7 +52,7 @@ public class CassandraMigrationAutoConfigurationTest {
         AnnotationConfigApplicationContext context =
                 new AnnotationConfigApplicationContext();
         addEnvironment(context, "cassandra.migration.keyspace-name:test_keyspace");
-        addEnvironment(context, "cassandra.migration.table-prefix:test_");
+        addEnvironment(context, "cassandra.migration.table-prefix:test");
         context.register(ClusterConfig.class, CassandraMigrationAutoConfiguration.class);
         context.refresh();
         Cluster cluster = context.getBean(Cluster.class);