+++ cassandra-migration/src/test/java/org/cognitor/cassandra/it/migration/DatabaseTest.java
@@ -1,6 +1,9 @@
 package org.cognitor.cassandra.it.migration;
 
-import com.datastax.driver.core.*;
+import com.datastax.driver.core.KeyspaceMetadata;
+import com.datastax.driver.core.Row;
+import com.datastax.driver.core.Session;
+import com.datastax.driver.core.SimpleStatement;
 import org.cognitor.cassandra.CassandraJUnitRule;
 import org.cognitor.cassandra.migration.Database;
 import org.cognitor.cassandra.migration.MigrationException;
@@ -8,7 +11,6 @@ import org.cognitor.cassandra.migration.MigrationRepository;
 import org.cognitor.cassandra.migration.MigrationTask;
 import org.cognitor.cassandra.migration.keyspace.Keyspace;
 import org.cognitor.cassandra.migration.keyspace.NetworkStrategy;
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
@@ -27,27 +29,48 @@ public class DatabaseTest {
     @Rule
     public final CassandraJUnitRule cassandra = new CassandraJUnitRule(DEFAULT_SCRIPT_LOCATION, "cassandra.yml");
 
-    private Database database;
-
-    @Before
-    public void setUp() {
-        database = new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE);
-    }
-
     @Test
     public void shouldReturnSchemaVersionOfZeroInAnEmptyDatabase() {
+        Database database = new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE);
         assertThat(database.getVersion(), is(equalTo(0)));
     }
 
     @Test
+    public void shouldCreateMigrationTableWithPrefixWhenPrefixGiven() {
+        new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE, "prefix");
+        assertThat(cassandra.getCluster().getMetadata().getKeyspace(CassandraJUnitRule.TEST_KEYSPACE)
+                .getTable("prefix_schema_migration"), is(notNullValue()));
+        assertThat(cassandra.getCluster().getMetadata().getKeyspace(CassandraJUnitRule.TEST_KEYSPACE)
+                .getTable("schema_migration"), is(nullValue()));
+    }
+
+    @Test
+    public void shouldApplyMigrationToDatabaseWhenMigrationsAndPrefixAndEmptyDatabaseGiven() {
+        final String prefix = "prefix";
+        Database database = new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE, prefix);
+        MigrationTask migrationTask = new MigrationTask(database, new MigrationRepository("cassandra/migrationtest/successful"));
+        migrationTask.migrate();
+        // after migration the database object is closed
+        database = new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE, prefix);
+        assertThat(database.getVersion(), is(equalTo(3)));
+
+        List<Row> results = loadMigrations(prefix);
+        assertThat(results.size(), is(equalTo(3)));
+        assertThat(results.get(0).getBool("applied_successful"), is(true));
+        assertThat(results.get(1).getBool("applied_successful"), is(true));
+        assertThat(results.get(2).getBool("applied_successful"), is(true));
+    }
+
+    @Test
     public void shouldApplyMigrationToDatabaseWhenMigrationsAndEmptyDatabaseGiven() {
+        Database database = new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE);
         MigrationTask migrationTask = new MigrationTask(database, new MigrationRepository("cassandra/migrationtest/successful"));
         migrationTask.migrate();
         // after migration the database object is closed
         database = new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE);
         assertThat(database.getVersion(), is(equalTo(3)));
 
-        List<Row> results = loadMigrations();
+        List<Row> results = loadMigrations("");
         assertThat(results.size(), is(equalTo(3)));
         assertThat(results.get(0).getBool("applied_successful"), is(true));
         assertThat(results.get(0).getTimestamp("executed_at"), is(not(nullValue())));
@@ -65,6 +88,7 @@ public class DatabaseTest {
 
     @Test
     public void shouldNotApplyAnyMigrationWhenDatabaseAndScriptsAreAtSameVersion() {
+        Database database = new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE);
         // provide a path without scripts to simulate this
         MigrationRepository repository = new MigrationRepository("migrationtest");
         new MigrationTask(database, repository).migrate();
@@ -74,6 +98,7 @@ public class DatabaseTest {
 
     @Test
     public void shouldThrowExceptionAndLogFailedMigrationWhenWrongMigrationScriptGiven() {
+        Database database = new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE);
         MigrationRepository repository = new MigrationRepository("cassandra/migrationtest/failing/brokenstatement");
         MigrationException exception = null;
         try {
@@ -86,7 +111,7 @@ public class DatabaseTest {
         assertThat(exception.getScriptName(), is(equalTo("001_init.cql")));
         assertThat(exception.getStatement(), is(equalTo("CREATE TABLE PERSON (id uuid primary key, name varcha);")));
 
-        List<Row> results = loadMigrations();
+        List<Row> results = loadMigrations("");
         assertThat(results.size(), is(equalTo(1)));
         assertThat(results.get(0).getBool("applied_successful"), is(false));
         assertThat(results.get(0).getTimestamp("executed_at"), is(not(nullValue())));
@@ -94,6 +119,7 @@ public class DatabaseTest {
 
     @Test
     public void shouldCreateKeyspaceWhenDatabaseWithoutKeyspaceAndKeyspaceDefinitionGiven() {
+        Database database = new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE);
         assertThat(cassandra.getCluster().getMetadata().getKeyspace("new_keyspace"), is(nullValue()));
         Keyspace keyspace = new Keyspace("new_keyspace");
         Database db = new Database(cassandra.getCluster(), keyspace);
@@ -108,6 +134,7 @@ public class DatabaseTest {
 
     @Test
     public void shouldCreateKeyspaceWhenDatabaseWithoutKeyspaceAndNetworkKeyspaceDefinitionGiven() {
+        Database database = new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE);
         assertThat(cassandra.getCluster().getMetadata().getKeyspace("network_keyspace"), is(nullValue()));
         Keyspace keyspace = new Keyspace("network_keyspace")
                 .with(new NetworkStrategy().with("dc1", 1));
@@ -122,6 +149,7 @@ public class DatabaseTest {
 
     @Test
     public void shouldCreateFunctionWhenMigrationScriptWithFunctionGiven() {
+        Database database = new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE);
         MigrationTask migrationTask = new MigrationTask(database, new MigrationRepository("cassandra/migrationtest/function"));
         migrationTask.migrate();
         database = new Database(cassandra.getCluster(), CassandraJUnitRule.TEST_KEYSPACE);
@@ -130,9 +158,13 @@ public class DatabaseTest {
                 .getKeyspace(CassandraJUnitRule.TEST_KEYSPACE).getFunctions().size(), is(equalTo(1)));
     }
 
-    private List<Row> loadMigrations() {
+    private List<Row> loadMigrations(String tablePrefix) {
         Session session = cassandra.getCluster().connect(CassandraJUnitRule.TEST_KEYSPACE);
-        ResultSet resultSet = session.execute(new SimpleStatement("SELECT * FROM schema_migration;"));
-        return resultSet.all();
+        if (tablePrefix == null || tablePrefix.isEmpty()) {
+            return session.execute(
+                    new SimpleStatement("SELECT * FROM schema_migration;")).all();
+        }
+        return session.execute(
+                new SimpleStatement(String.format("SELECT * FROM %s_schema_migration;", tablePrefix))).all();
     }
 }